+++ include/openrave/kinbody.h
@@ -53,6 +53,7 @@ enum GeometryType : uint8_t
     GT_Container=5, ///< a container shaped geometry that has inner and outer extents. container opens on +Z. The origin is at the bottom of the base.
     GT_Cage=6, ///< a container shaped geometry with removable side walls. The side walls can be on any of the four sides. The origin is at the bottom of the base. The inner volume of the cage is measured from the base to the highest wall.
     GT_CalibrationBoard=7, ///< a box shaped geometry with grid of cylindrical dots of two sizes. The dots are always on the +z side of the box and are oriented towards z-axis.
+    GT_Axial = 8, ///< a geometry defined by many slices along an axis, oriented towards z-axis
 };
 
 enum DynamicsConstraintsType : int8_t
@@ -355,6 +356,7 @@ public:
                    && _vGeomData3 == other._vGeomData3
                    && _vGeomData4 == other._vGeomData4
                    && _vSideWalls == other._vSideWalls
+                   && _vAxialSlices == other._vAxialSlices
                    && _vDiffuseColor == other._vDiffuseColor
                    && _vAmbientColor == other._vAmbientColor
                    && _meshcollision == other._meshcollision
@@ -572,6 +574,51 @@ public:
         /// \brief Generates the calibration board's dot grid mesh based on calibration board settings
         void GenerateCalibrationBoardDotMesh(TriMesh& tri, float fTessellation=1) const;
 
+
+        /// \brief A slice used to construct axial geometry.
+        struct AxialSlice
+        {
+            dReal zOffset; ///< z coordinate of the axial slice
+            dReal radius; ///< radius of the axial slice
+
+            int Compare(const AxialSlice& rhs, dReal fUnitScale=1.0, dReal fEpsilon=10e-7) const;
+            bool operator==(const AxialSlice& other) const {
+                return zOffset == other.zOffset && radius == other.radius;
+            }
+            bool operator!=(const AxialSlice& other) const {
+                return !operator==(other);
+            }
+            bool operator<(const AxialSlice& other) const {
+                return zOffset < other.zOffset;
+            }
+        };
+
+        /// \brief An axial geometry representation 
+        /// 
+        /// An axial might look like drawing below. 
+        /// The dotted lines represents the z and y axes. Each row 
+        /// represents an axial slice.
+        ///
+        /// There has to be at least two axial slices in an axial, 
+        /// which are called top and bottom axial slices. 
+        ///
+        /// If the radius of the top and bottom slices are the same, 
+        /// the resulting shape is essentially a cylinder.
+        /// 
+        /// The axial slices stored in this vector are not sorted by
+        /// the z coordinate.
+        ///
+        ///         z
+        ///         :
+        ///      ___:___
+        ///  ...|...:...|... y
+        ///     |   :   |
+        ///      \__:__/
+        ///         : 
+        ///         :
+        ///   
+        std::vector<AxialSlice> _vAxialSlices;
+
         enum GeometryInfoField : uint32_t
         {
             GIF_Transform = (1 << 0), // _t field
@@ -717,6 +764,9 @@ public:
         }
         inline const Vector& GetPositiveCropContainerEmptyMargins() const {
             return _info._vPositiveCropContainerEmptyMargins;
+        }
+        inline int GetNumberOfAxialSlices() const {
+            return _info._vAxialSlices.size();
         }
 
         /// \brief returns the local collision mesh