+++ plugins/fclrave/fclspace.cpp
@@ -459,6 +459,7 @@ CollisionGeometryPtr FCLSpace::_CreateFCLGeomFromGeometryInfo(const KinBody::Geo
             return std::make_shared<fcl::Cylinder>(info._vGeomData.x, info._vGeomData.y);
         }
 
+    case OpenRAVE::GT_Axial:
     case OpenRAVE::GT_Container:
     case OpenRAVE::GT_TriMesh:
     case OpenRAVE::GT_Cage: