+++ plugins/qtosgrave/osgrenderitem.cpp
@@ -446,6 +446,7 @@ void KinBodyItem::Load()
                     }
                 }
                 //  Extract geometry from collision Mesh
+                case GT_Axial:
                 case GT_Cage:
                 case GT_Container:
                 case GT_TriMesh: {
@@ -475,7 +476,7 @@ void KinBodyItem::Load()
                     geode->addDrawable(geom);
                     pgeometrydata->addChild(geode);
 
-                    if(orgeom->GetType() == GT_TriMesh){
+                    if(orgeom->GetType() == GT_TriMesh || orgeom->GetType() == GT_Axial){
                         // CropContainerMargins and CropContainerEmptyMargins only exists in GT_Cage and GT_Container
                         break;
                     }