+++ src/libopenrave-core/jsonparser/jsonreader.cpp
@@ -226,7 +226,7 @@ public:
         }
 
         // set global scale when initalize jsonreader.
-        _fGlobalScale = 1.0 / _penv->GetUnit().second;
+        _fGlobalScale = GetLengthUnitStandardValue<dReal>(_penv->GetUnitInfo().lengthUnit);
         _deserializeOptions = 0;
 
 
@@ -804,6 +804,11 @@ protected:
 
     inline dReal _GetUnitScale(const rapidjson::Value& doc, dReal defaultScale)
     {
+        if (doc.HasMember("unitInfo")) {
+            UnitInfo unitInfo;
+            orjson::LoadJsonValueByKey(doc, "unitInfo", unitInfo);
+            return 1.0 / GetLengthUnitStandardValue<dReal>(unitInfo.lengthUnit) * _fGlobalScale * _fGeomScale;
+        }
         std::pair<std::string, dReal> unit = {"", defaultScale};
         orjson::LoadJsonValueByKey(doc, "unit", unit);
         return unit.second * _fGlobalScale * _fGeomScale;